requirements.txt
blinker==1.6.2
cachelib==0.10.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.7
cryptography==39.0.0
Flask==2.3.2
Flask-Caching==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pycparser==2.21
PyJWT==2.6.0
python-dotenv==0.21.0
requests==2.28.1
urllib3==1.26.16
Werkzeug==2.3.7